+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changes:
 - Simplify font sizing configuration
 - Several small UI fixes
 - Use letter icon for shortcuts too
+- Always show all barcode types in manual entry
+- Revert to valid barcode type if the value is not valid for the selected barcode type in edit activity
 
 ## v1.12 (2021-03-30)